llama : merge gguf-util.h in llama.cpp
